Moving unit tests into modules
Removing unneeded variable
