+++ rusty-machine/tests/linalg/mat.rs
@@ -1,76 +1,4 @@
 use rm::linalg::matrix::Matrix;
-use rm::linalg::vector::Vector;
-use libnum::abs;
-
-#[test]
-fn create_mat_zeros() {
-    let a = Matrix::<f32>::zeros(10, 10);
-
-    assert_eq!(a.rows(), 10);
-    assert_eq!(a.cols(), 10);
-
-    for i in 0..10 {
-        for j in 0..10 {
-            assert_eq!(a[[i, j]], 0.0);
-        }
-    }
-}
-
-#[test]
-fn create_mat_identity() {
-    let a = Matrix::<f32>::identity(4);
-
-    assert_eq!(a.rows(), 4);
-    assert_eq!(a.cols(), 4);
-
-    assert_eq!(a[[0, 0]], 1.0);
-    assert_eq!(a[[1, 1]], 1.0);
-    assert_eq!(a[[2, 2]], 1.0);
-    assert_eq!(a[[3, 3]], 1.0);
-
-    assert_eq!(a[[0, 1]], 0.0);
-    assert_eq!(a[[2, 1]], 0.0);
-    assert_eq!(a[[3, 0]], 0.0);
-}
-
-#[test]
-fn create_mat_diag() {
-    let a = Matrix::from_diag(&[1.0, 2.0, 3.0, 4.0]);
-
-    assert_eq!(a.rows(), 4);
-    assert_eq!(a.cols(), 4);
-
-    assert_eq!(a[[0, 0]], 1.0);
-    assert_eq!(a[[1, 1]], 2.0);
-    assert_eq!(a[[2, 2]], 3.0);
-    assert_eq!(a[[3, 3]], 4.0);
-
-    assert_eq!(a[[0, 1]], 0.0);
-    assert_eq!(a[[2, 1]], 0.0);
-    assert_eq!(a[[3, 0]], 0.0);
-}
-
-#[test]
-fn transpose_mat() {
-    let a = Matrix::new(5, 2, vec![1., 2., 3., 4., 5., 6., 7., 8., 9., 10.]);
-
-    let c = a.transpose();
-
-    assert_eq!(c.cols(), a.rows());
-    assert_eq!(c.rows(), a.cols());
-
-    assert_eq!(a[[0, 0]], c[[0, 0]]);
-    assert_eq!(a[[1, 0]], c[[0, 1]]);
-    assert_eq!(a[[2, 0]], c[[0, 2]]);
-    assert_eq!(a[[3, 0]], c[[0, 3]]);
-    assert_eq!(a[[4, 0]], c[[0, 4]]);
-    assert_eq!(a[[0, 1]], c[[1, 0]]);
-    assert_eq!(a[[1, 1]], c[[1, 1]]);
-    assert_eq!(a[[2, 1]], c[[1, 2]]);
-    assert_eq!(a[[3, 1]], c[[1, 3]]);
-    assert_eq!(a[[4, 1]], c[[1, 4]]);
-
-}
 
 #[test]
 fn matrix_lup_decomp() {
@@ -97,58 +25,6 @@ fn matrix_lup_decomp() {
     for i in 0..25 {
         assert_eq!(e.data()[i], k.data()[i]);
     }
-}
-
-#[test]
-fn matrix_diag() {
-    let a = Matrix::new(3, 3, vec![1., 3., 5., 2., 4., 7., 1., 1., 0.]);
-
-    let b = a.is_diag();
-
-    assert!(!b);
-
-    let c = Matrix::new(3, 3, vec![1., 0., 0., 0., 2., 0., 0., 0., 3.]);
-    let d = c.is_diag();
-
-    assert!(d);
-}
-
-#[test]
-fn matrix_det() {
-    let a = Matrix::new(2, 2, vec![2., 3., 1., 2.]);
-    let b = a.det();
-
-    assert_eq!(b, 1.);
-
-    let c = Matrix::new(3, 3, vec![1., 2., 3., 4., 5., 6., 7., 8., 9.]);
-    let d = c.det();
-
-    assert_eq!(d, 0.);
-
-    let e = Matrix::<f64>::new(5,
-                               5,
-                               vec![1., 2., 3., 4., 5., 3., 0., 4., 5., 6., 2., 1., 2., 3., 4.,
-                                    0., 0., 0., 6., 5., 0., 0., 0., 5., 6.]);
-
-    let f = e.det();
-
-    println!("det is {0}", f);
-    let error = abs(f - 99.);
-    assert!(error < 1e-10);
-}
-
-#[test]
-fn matrix_solve() {
-    let a = Matrix::new(2, 2, vec![2., 3., 1., 2.]);
-
-    let y = Vector::new(vec![8., 5.]);
-
-    let x = a.solve(y);
-
-    assert_eq!(x.size(), 2);
-
-    assert_eq!(x[0], 1.);
-    assert_eq!(x[1], 2.);
 }
 
 #[test]
